requirements.txt
torch_geometric
numpy
networkx
jax
scikit-learn